requirements.txt
numpy
opencv_python
opencv_python_headless
Roboflow
supervision
Pillow
PyAutoGUI
pynput
pytesseract
threading
python-Levenshtein
PyQt5
ultralytics